simple gui for wdpassport-utils!
This is a first attempt of a simple gui for the wdpassport-utils
code to make it a little bit more convenient for the average user to use.

It was my first try in python programing, so it might be crappy,
ugly and unprofessional. But nevertheless it worked for me.

tested on Ubuntu 13.10 x86_64